+++ boosty_downloader/src/infrastructure/boosty_api/models/post/post_data_types/__init__.py
@@ -6,6 +6,7 @@ from .post_data_link import BoostyPostDataLinkDTO
 from .post_data_list import BoostyPostDataListDTO
 from .post_data_ok_video import BoostyPostDataOkVideoDTO
 from .post_data_text import BoostyPostDataTextDTO
+from .post_data_unknown import BoostyPostDataUnknownDTO
 from .post_data_video import BoostyPostDataExternalVideoDTO
 
 __all__ = [
@@ -18,4 +19,5 @@ __all__ = [
     'BoostyPostDataListDTO',
     'BoostyPostDataOkVideoDTO',
     'BoostyPostDataTextDTO',
+    'BoostyPostDataUnknownDTO',
 ]